Add npm-run as alias for npm-run-all
Editing tasks from node'ran to npm-run-all'ran tasks it is much more convenient to do if it only demands to add a `-` compared to also adding the last `-all` part. It also looks better (in my opinion). 

One argument against this is that it becomes harder for people to recognise that this command is not the native command and how to find the documentation. Having the choice does not limit the ones who want to keep the original notation and as the package-unrelated command names `run-s` and `run-p` are already supported, we are not introducing added complexity.